Backend/requirements.txt
FastAPI
Uvicorn
SQLModel
PyJwt
python-dotenv
psycopg2
passlib[bcrypt]
